pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "supergraph"
version = "0.1.0"
description = "JSON Query DSL for microservices - alternative to GraphQL Federation"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Supergraph Authors" }
]
maintainers = [
    { name = "Supergraph Authors" }
]
keywords = [
    "microservices",
    "api",
    "gateway",
    "graphql-alternative",
    "json-query",
    "fastapi",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
]
service = [
    "uvicorn>=0.23.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "alembic>=1.12.0",
]
all = [
    "supergraph[dev,service]",
]

[project.scripts]
supergraph = "supergraph.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/supergraph"
Documentation = "https://github.com/your-org/supergraph#readme"
Repository = "https://github.com/your-org/supergraph"
Issues = "https://github.com/your-org/supergraph/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/supergraph"]

# Include bundled playground frontend (run scripts/build_playground.sh first)
[tool.hatch.build.targets.wheel.shared-data]
"src/supergraph/playground/dist" = "supergraph/playground/dist"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/playground",
    "/scripts",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["supergraph"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
